More fixes to issues identified by Sonar

+++ tools/gcc-bridge/compiler/src/main/java/org/renjin/gcc/gimple/GimpleFunction.java
@@ -119,7 +119,7 @@ public class GimpleFunction implements GimpleDecl {
   public GimpleVarDecl addVarDecl(GimpleType type) {
     // find unused id
     Set<Long> usedIds = usedIds();
-    int newId = 1000;
+    long newId = 1000;
     while(usedIds.contains(newId)) {
       newId++;
     }
